+++ Net_Core/Web.Angular/projects/gw-lib/core/configuration/src/db-information.model.ts
@@ -1,12 +1,12 @@
 export interface IDBInformation {
+    databaseInformationSeqId: number;
     enableInheritance: number;
-    informationSeqId: number;
     version: string;
 }
 
 export class DBInformation implements IDBInformation {
+	public databaseInformationSeqId: number = -1;
 	public enableInheritance: number = -1;
-	public informationSeqId: number = -1;
 	public version: string = '';
 
 	constructor() {}